Created Movies Scrapper
Welcome to our GitHub repository! This is a project that aims to get Hollywood Movies data by web scraping.

Our repository contains all the necessary code required to implement our solution, this code is organized, and easy to navigate and understand.

We have used Python to build our solution, and we have included detailed comments and documentation to help you understand how everything works.

In the "docs" folder, you will find all the necessary documentation to get started with our solution. This includes installation instructions. We have also included a detailed README file to help you get started with the repository.

We believe that collaboration is key to success, which is why we have made our repository open-source. We welcome contributions from the community, and we encourage you to get involved by submitting issues, feature requests, or pull requests.

Thank you for your interest in our project, and we hope that our solution can help you solve your problem!